% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truedriver.R
\name{as.GraphParameter}
\alias{as.GraphParameter}
\title{Combines graph parameters into a list.}
\usage{
as.GraphParameter(xBreaks = c(0, 1), yBreaks = c(-1, 1),
  xLim = c(min(c(xBreaks, -0.1)), max(xBreaks)), yLim = c(min(yBreaks),
  max(yBreaks)), xLabel = "Relative importance",
  yLabel = "Net satisfaction", xlogScale = FALSE)
}
\arguments{
\item{xBreaks}{vector with breaks for x axis}

\item{yBreaks}{vector with breaks for y axis}

\item{xLim}{vector with limits for coord_cartesian x-axis}

\item{yLim}{vector with limits for coord_cartesian x-axis}

\item{xLabel}{x axis label}

\item{yLabel}{y axis label}

\item{xlogScale}{If TRUE will plot x-axis on log scale}
}
\description{
Combines graph parameters into a list.
}
