% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truedriver.R
\name{combine_truedriver}
\alias{combine_truedriver}
\title{Combines result from two true driver analyses.}
\usage{
combine_truedriver(td1, td2, segments)
}
\arguments{
\item{td1}{A dataframe}

\item{td2}{A data frame}

\item{segments}{Segments}
}
\value{
A list
}
\description{
This combines the result of two separate true drivera analyses into 
a single dataframe, to plot using \link{plot_truedriver_segmental}
}
