% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truedriver.R
\name{netScore}
\alias{netScore}
\title{Calculates a net score.}
\usage{
netScore(x)
}
\arguments{
\item{x}{An ordered factor}
}
\description{
This assumes that x is an ordered factor, from low to high
It then calculates a net percentage score
}
