% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truedriver.R
\name{plot_truedriver_segmental}
\alias{plot_truedriver_segmental}
\title{Plots segmental true driver graph.}
\usage{
plot_truedriver_segmental(impsats, GraphParameters,
  meansFunction = c("median", "mean"), means = NULL, facet = FALSE, ...)
}
\arguments{
\item{impsats}{A dataframe}

\item{GraphParameters}{Result of \code{\link{as.GraphParameter}}}

\item{meansFunction}{Vector of length 2, containing functions to use for calculating midpoints of importance and satisfaction. Defaults to median and mean.}

\item{means}{Numeric vector of length 2. First element is mean of importance, second element is mean of satisfaction. If null, defaults to calculated means of data}

\item{facet}{Logical. If TRUE, creates plot with a facet for each segment, otherwise creates single facet with lines connecting observations.}

\item{...}{Passed to ggplot opts}
}
\value{
A list
}
\description{
This is a thin wrapper around relaimp(), returning a list with results and key indicators
}
