% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truedriver-package.R
\docType{package}
\name{truedriver-package}
\alias{truedriver-package}
\title{Tools to perform True Driver Analysis.}
\description{
Tools to perform True Driver Analysis.
}
\details{
Provides wrappers around the \code{relaimpo} package

\itemize{
\item True driver analys: \code{\link{truedriver}} 
}
}
\seealso{
\code{\link{truedriver}}
}
\author{
Andrie de Vries \email{andrie.de.vries@pentalibra.com}
}
\keyword{package}
