% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truedriver.R
\name{truedriver}
\alias{truedriver}
\title{Performs true driver analysis.}
\usage{
truedriver(model, data, impLabels = labels(terms(model, data = data)),
  type = c("pratt", "pmvd"), meansFunction = c("median", "mean"))
}
\arguments{
\item{model}{A model formula}

\item{data}{A dataframe}

\item{impLabels}{Character vector with labels for importance elements. Defaults to}

\item{type}{Passed to \code{\link[relaimpo]{calc.relimp}}}

\item{meansFunction}{Vector of length 2, containing functions to use for calculating midpoints of importance and satisfaction. Defaults to median and mean.}
}
\value{
A list:
\item{vinflation}{Variance inflation factor, see \code{\link[car]{vif}}}
\item{impsats}{A data frame with columns for labels, imp (importance) and sats (satisfaction)}
\item{means}{A list with two elements: \code{means} are the midpoints of imp and sats, while \code{meansFunction} contain the functions used to calculate the midpoint (typically \code{median} and \code{mean})}
}
\description{
This is a thin wrapper around relaimp(), returning a list with results and key indicators.
}
\details{
By default, uses the pratt method, which is fast but can yield negative coefficients. The pmvd method (proportionate marginal variance decomposition) is better but expoentially slower depending on number of covariates.
}
\examples{
data(diamonds, package="ggplot2")
x <- truedriver(price~cut+color+carat+clarity, data=diamonds, type="pmvd")
plot(x)

x <- truedriver(price~., data=diamonds[, c("price", "cut", "color", "clarity", "carat")])
plot(x)
}
